refactor: extract task-scheduling.ts from daemon/main.ts
Move dynamic task scheduling (definition loading, last-run seeding,
ScheduledJobContext with pre-conditions and notifications, job
registration) into registerScheduledTasks().
